requirements.txt
flask==2.0.2
nltk==3.6.2
spacy==3.0.6
python-dotenv==0.19.1
flask-jwt-extended==4.3.1
